requirements.txt
langchain_chroma==0.2.4
langchain_core==0.3.62
langchain_huggingface==0.2.0
langchain_openai==0.3.18
langgraph==0.4.7
pandas==2.2.3
pydantic==2.11.5
pytest==8.3.5
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.45.1
streamlit_lottie==0.0.5
tqdm==4.67.1
typing_extensions==4.13.2
pytest-json-report==1.5.0
redis==6.2.0
gdown==5.2.0